Refactor 1 - Prevented future technical debts(Hard coded details shifted to env, added versioning in APIs
